Updated Requirements.txt
