+++ requirements.txt
@@ -1,31 +1,41 @@
 async-generator==1.10
-attrs==21.4.0
-certifi==2022.5.18.1
-cffi==1.15.0
-charset-normalizer==2.0.12
-cryptography==37.0.2
-h11==0.13.0
-idna==3.3
+attrs==22.1.0
+autopep8==1.7.0
+certifi==2022.9.24
+cffi==1.15.1
+charset-normalizer==2.1.1
+colorama==0.4.5
+cryptography==38.0.1
+exceptiongroup==1.0.0rc9
+h11==0.14.0
+idna==3.4
 oauthlib==3.2.1
-outcome==1.1.0
+outcome==1.2.0
+packaging==21.3
+pip-review==1.2.0
 praw==7.6.0
 prawcore==2.3.0
+pycodestyle==2.9.1
 pycparser==2.21
 PyMsgBox==1.0.9
-pyOpenSSL==22.0.0
+pyOpenSSL==22.1.0
+pyparsing==3.0.9
 pyperclip==1.8.2
 PySocks==1.7.1
-python-dotenv==0.20.0
-requests==2.27.1
+python-dotenv==0.21.0
+redvid==1.2.0
+requests==2.28.1
 requests-oauthlib==1.3.1
-selenium==4.1.5
-sniffio==1.2.0
+selenium==4.5.0
+sniffio==1.3.0
 sortedcontainers==2.4.0
-trio==0.20.0
+toml==0.10.2
+tqdm==4.64.1
+trio==0.22.0
 trio-websocket==0.9.2
-tweepy==4.10.0
+tweepy==4.10.1
 update-checker==0.18.0
-urllib3==1.26.9
-webdriver-manager==3.7.0
-websocket-client==1.3.2
-wsproto==1.1.0
+urllib3==1.26.12
+webdriver-manager==3.8.3
+websocket-client==1.4.1
+wsproto==1.2.0
